Q2: Botclip support through AUX, better content flag mapping
Quake 2 brush flags are filtered to ensure no collisions with Quake 3
ones (CURRENT_UP was counting as BOTCLIP, for example).

If CONTENTFLAGS_AUX is set on a Quake 2 brush, it's marked as BOTCLIP.

AAS creation process now uses separate header with common flags to
make sure we're not mixing up Quake 2 and Quake 3 ones.